refactor(users-info): split tests_security_scoping into 4 focused test modules
Split the 1145-line monolithic test file into category-based modules:
- tests_tenant_scoping: tenant isolation, CRUD, transaction smoke tests
- tests_resource_scoping: owner/city constraint enforcement
- tests_pdp_deny: explicit PDP denial scenarios across all endpoints
- tests_pdp_error: PDP internal error propagation
