--- examples/modkit/users-info/users-info/src/domain/service/tests_security_scoping.rs
@@ -1,1145 +1,0 @@
-#![allow(clippy::unwrap_used, clippy::expect_used)]
-
-use uuid::Uuid;
-
-use std::sync::Arc;
-
-use crate::domain::error::DomainError;
-use crate::domain::service::ServiceConfig;
-use crate::test_support::{
-    DenyAllAuthZResolver, FailingAuthZResolver, OwnerCityAuthZResolver, build_services,
-    build_services_with_authz, ctx_allow_tenants, ctx_deny_all, ctx_for_subject, inmem_db,
-    seed_user,
-};
-use modkit_db::DBProvider;
-use users_info_sdk::{NewAddress, NewCity, NewUser};
-
-#[tokio::test]
-async fn tenant_scope_only_sees_its_tenant() {
-    let db = inmem_db().await;
-    let tenant1 = Uuid::new_v4();
-    let tenant2 = Uuid::new_v4();
-
-    let user1 = Uuid::new_v4();
-    let user2 = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user1, tenant1, "u1@example.com", "U1").await;
-    seed_user(&conn, user2, tenant2, "u2@example.com", "U2").await;
-
-    let services = build_services(db.clone(), ServiceConfig::default());
-    let ctx_t1 = ctx_allow_tenants(&[tenant1]);
-
-    let page = services
-        .users
-        .list_users_page(&ctx_t1, &modkit_odata::ODataQuery::default())
-        .await
-        .unwrap();
-    assert_eq!(page.items.len(), 1);
-    assert_eq!(page.items[0].tenant_id, tenant1);
-}
-
-#[tokio::test]
-async fn deny_all_returns_forbidden() {
-    let db = inmem_db().await;
-    let tenant = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, Uuid::new_v4(), tenant, "u@example.com", "U").await;
-
-    let services = build_services(db.clone(), ServiceConfig::default());
-    let ctx = ctx_deny_all();
-
-    // Anonymous context has no tenant → mock returns empty constraints
-    // → Decision Matrix: require_constraints=true + empty → ConstraintsRequiredButAbsent → Forbidden
-    let result = services
-        .users
-        .list_users_page(&ctx, &modkit_odata::ODataQuery::default())
-        .await;
-    let err = result.unwrap_err();
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for anonymous context, got: {err:?}"
-    );
-}
-
-#[tokio::test]
-async fn create_user_with_transaction() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-
-    let services = build_services(db.clone(), ServiceConfig::default());
-    // Use a context with tenants, not root, because insert requires tenant scope
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let new_user = NewUser {
-        id: None,
-        tenant_id,
-        email: "test@example.com".to_owned(),
-        display_name: "Test User".to_owned(),
-    };
-
-    let result = services.users.create_user(&ctx, new_user).await;
-    assert!(result.is_ok(), "create_user failed: {:?}", result.err());
-
-    let created = result.unwrap();
-    assert_eq!(created.email, "test@example.com");
-    assert_eq!(created.display_name, "Test User");
-    assert_eq!(created.tenant_id, tenant_id);
-}
-
-#[tokio::test]
-async fn dbprovider_transaction_smoke() {
-    use crate::infra::storage::entity::user::{ActiveModel, Entity as UserEntity};
-    use modkit_db::secure::{AccessScope, secure_insert};
-    use sea_orm::Set;
-    use time::OffsetDateTime;
-
-    let db = inmem_db().await;
-    let provider: DBProvider<modkit_db::DbError> = DBProvider::new(db.clone());
-
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let now = OffsetDateTime::now_utc();
-    let scope = AccessScope::for_tenants(vec![tenant_id]);
-
-    provider
-        .transaction(|tx| {
-            Box::pin(async move {
-                let user = ActiveModel {
-                    id: Set(user_id),
-                    tenant_id: Set(tenant_id),
-                    email: Set("tx@example.com".to_owned()),
-                    display_name: Set("Tx User".to_owned()),
-                    created_at: Set(now),
-                    updated_at: Set(now),
-                };
-                let _ = secure_insert::<UserEntity>(user, &scope, tx).await?;
-                Ok(())
-            })
-        })
-        .await
-        .unwrap();
-}
-
-#[tokio::test]
-async fn create_address_validates_user_exists() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-
-    let services = build_services(db.clone(), ServiceConfig::default());
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let bogus_user_id = Uuid::new_v4();
-    let result = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id: bogus_user_id,
-                city_id: Uuid::new_v4(),
-                street: "Nowhere St".to_owned(),
-                postal_code: "00000".to_owned(),
-            },
-        )
-        .await;
-
-    assert!(result.is_err(), "Expected error for non-existent user");
-}
-
-#[tokio::test]
-async fn create_address_forces_user_tenant() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "addr@example.com", "Addr User").await;
-
-    let services = build_services(db.clone(), ServiceConfig::default());
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    // Create a city in the same tenant
-    let city = services
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Test City".to_owned(),
-                country: "TC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Pass a different tenant_id in NewAddress — the service must override it
-    let different_tenant = Uuid::new_v4();
-    let created = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id: different_tenant,
-                user_id,
-                city_id: city.id,
-                street: "123 Main St".to_owned(),
-                postal_code: "12345".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(
-        created.tenant_id, tenant_id,
-        "Address tenant must match user's tenant, not the request"
-    );
-}
-
-#[tokio::test]
-async fn put_user_address_creates_then_updates() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "put@example.com", "Put User").await;
-
-    let services = build_services(db.clone(), ServiceConfig::default());
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let city = services
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "City A".to_owned(),
-                country: "CA".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // First PUT — should create
-    let created = services
-        .addresses
-        .put_user_address(
-            &ctx,
-            user_id,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "First St".to_owned(),
-                postal_code: "11111".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(created.street, "First St");
-    assert_eq!(created.tenant_id, tenant_id);
-
-    // Second PUT — should update
-    let updated = services
-        .addresses
-        .put_user_address(
-            &ctx,
-            user_id,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "Second St".to_owned(),
-                postal_code: "22222".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(updated.id, created.id, "Should update the same address");
-    assert_eq!(updated.street, "Second St");
-    assert_eq!(updated.postal_code, "22222");
-}
-
-// ---------------------------------------------------------------------------
-// Owner + City authorization tests (using OwnerCityAuthZResolver)
-// ---------------------------------------------------------------------------
-
-/// User A creates an address for themselves — succeeds because
-/// `OwnerCityAuthZResolver` returns `eq(owner_id, subject.id)` and the
-/// `subject_id` matches the address's `user_id`.
-#[tokio::test]
-async fn owner_scope_allows_own_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "owner@example.com", "Owner").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(OwnerCityAuthZResolver),
-    );
-
-    // subject_id == user_id → PDP returns eq(owner_id, user_id) → matches
-    let ctx = ctx_for_subject(user_id, tenant_id);
-
-    let city = services
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Own City".to_owned(),
-                country: "OC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    let result = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "My Street".to_owned(),
-                postal_code: "11111".to_owned(),
-            },
-        )
-        .await;
-
-    assert!(result.is_ok(), "Owner should be able to create own address");
-    assert_eq!(result.unwrap().user_id, user_id);
-}
-
-/// User A tries to delete user B's address — fails because
-/// `OwnerCityAuthZResolver` returns `eq(owner_id, A)` but the address row
-/// has `user_id = B`, so the scoped DELETE matches 0 rows → `NotFound`.
-///
-/// Note: `secure_insert` only validates `tenant_id` on INSERT (there's no
-/// existing row to scope against). Owner/city constraints are enforced on
-/// reads, updates, and deletes via `.secure().scope_with()` WHERE clauses.
-#[tokio::test]
-async fn owner_scope_prevents_mutating_another_users_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_a = Uuid::new_v4();
-    let user_b = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_a, tenant_id, "a@example.com", "User A").await;
-    seed_user(&conn, user_b, tenant_id, "b@example.com", "User B").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(OwnerCityAuthZResolver),
-    );
-
-    let ctx_b = ctx_for_subject(user_b, tenant_id);
-
-    let city = services
-        .cities
-        .create_city(
-            &ctx_b,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Shared City".to_owned(),
-                country: "SC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // User B creates their own address (succeeds)
-    let addr_b = services
-        .addresses
-        .create_address(
-            &ctx_b,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id: user_b,
-                city_id: city.id,
-                street: "B Street".to_owned(),
-                postal_code: "22222".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // User A tries to delete user B's address
-    // PDP returns eq(owner_id, user_a) → scoped query adds WHERE user_id = user_a
-    // → address belongs to user_b → 0 rows matched → NotFound
-    let ctx_a = ctx_for_subject(user_a, tenant_id);
-    let delete_result = services.addresses.delete_address(&ctx_a, addr_b.id).await;
-
-    assert!(
-        delete_result.is_err(),
-        "User A must not be able to delete user B's address"
-    );
-
-    // User A tries to update user B's address via put_user_address
-    // The prefetch loads the address (allow_all), but PDP returns eq(owner_id, user_a)
-    // → scoped UPDATE adds WHERE user_id = user_a → 0 rows affected → error
-    let update_result = services
-        .addresses
-        .put_user_address(
-            &ctx_a,
-            user_b,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id: user_b,
-                city_id: city.id,
-                street: "Hacked St".to_owned(),
-                postal_code: "99999".to_owned(),
-            },
-        )
-        .await;
-
-    assert!(
-        update_result.is_err(),
-        "User A must not be able to update user B's address"
-    );
-
-    // Verify user B's address is untouched
-    let addr = services
-        .addresses
-        .get_user_address(&ctx_b, user_b)
-        .await
-        .unwrap()
-        .unwrap();
-    assert_eq!(addr.street, "B Street", "Address must remain unchanged");
-}
-
-/// User A tries to CREATE an address for user B — fails because
-/// `secure_insert` now validates all scope properties (not just `tenant_id`).
-/// PDP returns `eq(owner_id, user_a)` but the INSERT has `user_id = user_b`,
-/// so `validate_insert_scope` rejects the mismatch.
-#[tokio::test]
-async fn owner_scope_prevents_creating_address_for_another_user() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_a = Uuid::new_v4();
-    let user_b = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_a, tenant_id, "a@example.com", "User A").await;
-    seed_user(&conn, user_b, tenant_id, "b@example.com", "User B").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(OwnerCityAuthZResolver),
-    );
-
-    // subject is user_a → PDP returns eq(owner_id, user_a)
-    let ctx = ctx_for_subject(user_a, tenant_id);
-
-    let city = services
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Shared City".to_owned(),
-                country: "SC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Try to create address for user_b while authenticated as user_a
-    let result = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id: user_b,
-                city_id: city.id,
-                street: "Sneaky St".to_owned(),
-                postal_code: "99999".to_owned(),
-            },
-        )
-        .await;
-
-    assert!(
-        result.is_err(),
-        "User A must not be able to create an address for user B"
-    );
-}
-
-/// User creates an address in `city_1` (allowed) — succeeds.
-/// Then tries to update it to `city_2` — fails because PDP returns
-/// `eq(city_id, city_2)` but the scope constraint doesn't match the
-/// existing record's `city_id` during the scoped re-read.
-#[tokio::test]
-async fn city_scope_restricts_address_to_allowed_city() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "city@example.com", "City User").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(OwnerCityAuthZResolver),
-    );
-
-    let ctx = ctx_for_subject(user_id, tenant_id);
-
-    let city_1 = services
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Allowed City".to_owned(),
-                country: "AC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    let city_2 = services
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Forbidden City".to_owned(),
-                country: "FC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Create address in city_1 — PDP echoes eq(city_id, city_1) → matches INSERT
-    let created = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city_1.id,
-                street: "Good St".to_owned(),
-                postal_code: "11111".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(created.city_id, city_1.id);
-
-    // Now try to update the address to city_2.
-    // The update_address method prefetches the existing address (city_1),
-    // sends city_1 as resource property to PDP, PDP returns eq(city_id, city_1).
-    // The scoped re-read succeeds (existing record has city_1).
-    // But the final UPDATE writes city_2 into the row — the scope constraint
-    // eq(city_id, city_1) is applied in the UPDATE WHERE clause, which still
-    // matches the row (scope is checked against the existing row, not the new values).
-    //
-    // To demonstrate city restriction on CREATE (the cleaner scenario):
-    // Try creating a second address for a different user in city_2.
-    let user_id_2 = Uuid::new_v4();
-    seed_user(
-        &conn,
-        user_id_2,
-        tenant_id,
-        "city2@example.com",
-        "City User 2",
-    )
-    .await;
-    let ctx_2 = ctx_for_subject(user_id_2, tenant_id);
-
-    // OwnerCityAuthZResolver echoes back the city_id from resource properties.
-    // For city_2, it returns eq(city_id, city_2). secure_insert checks that
-    // the INSERT's city_id matches the constraint — it does, so this succeeds.
-    let created_2 = services
-        .addresses
-        .create_address(
-            &ctx_2,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id: user_id_2,
-                city_id: city_2.id,
-                street: "Other St".to_owned(),
-                postal_code: "22222".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(created_2.city_id, city_2.id);
-
-    // Verify that delete also respects owner scope: user_2 cannot delete user_1's address
-    let delete_result = services.addresses.delete_address(&ctx_2, created.id).await;
-
-    assert!(
-        delete_result.is_err(),
-        "User 2 must not be able to delete user 1's address (owner scope)"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Explicit PDP denial tests (decision=false → EnforcerError::Denied → Forbidden)
-// ---------------------------------------------------------------------------
-
-/// When the PDP explicitly denies access, `list_users_page` must return
-/// `DomainError::Forbidden` (not `InternalError` or any other variant).
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_list_users() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, Uuid::new_v4(), tenant_id, "u@example.com", "U").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .users
-        .list_users_page(&ctx, &modkit_odata::ODataQuery::default())
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden from explicit PDP denial, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `get_user` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_get_user() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "get@example.com", "Get User").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services.users.get_user(&ctx, user_id).await.unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for get_user, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `create_user` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_create_user() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .users
-        .create_user(
-            &ctx,
-            NewUser {
-                id: None,
-                tenant_id,
-                email: "new@example.com".to_owned(),
-                display_name: "New User".to_owned(),
-            },
-        )
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for create_user, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `update_user` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_update_user() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "upd@example.com", "Upd User").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .users
-        .update_user(
-            &ctx,
-            user_id,
-            users_info_sdk::UserPatch {
-                email: Some("updated@example.com".to_owned()),
-                display_name: None,
-            },
-        )
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for update_user, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `delete_user` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_delete_user() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "del@example.com", "Del User").await;
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services.users.delete_user(&ctx, user_id).await.unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for delete_user, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `list_addresses_page` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_list_addresses() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .addresses
-        .list_addresses_page(&ctx, &modkit_odata::ODataQuery::default())
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for list_addresses, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `get_address` → `DomainError::Forbidden`.
-/// The service prefetches the address (`allow_all`), then calls the enforcer
-/// which must propagate the denial.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_get_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(
-        &conn,
-        user_id,
-        tenant_id,
-        "addr-get@example.com",
-        "Addr Get",
-    )
-    .await;
-
-    // First create the address with a permissive resolver
-    let permissive = build_services(db.clone(), ServiceConfig::default());
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let city = permissive
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Deny City".to_owned(),
-                country: "DC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    let addr = permissive
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "Deny St".to_owned(),
-                postal_code: "00000".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Now use the denying resolver
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-
-    let err = services
-        .addresses
-        .get_address(&ctx, addr.id)
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for get_address, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `create_address` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_create_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "addr-cr@example.com", "Addr Cr").await;
-
-    // Create a city with a permissive resolver first
-    let permissive = build_services(db.clone(), ServiceConfig::default());
-    let pctx = ctx_allow_tenants(&[tenant_id]);
-
-    let city = permissive
-        .cities
-        .create_city(
-            &pctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Cr City".to_owned(),
-                country: "CC".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Now use the denying resolver
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "Denied St".to_owned(),
-                postal_code: "11111".to_owned(),
-            },
-        )
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for create_address, got: {err:?}"
-    );
-}
-
-/// Explicit PDP denial on `delete_address` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn pdp_denied_returns_forbidden_for_delete_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(
-        &conn,
-        user_id,
-        tenant_id,
-        "addr-del@example.com",
-        "Addr Del",
-    )
-    .await;
-
-    // Create address with permissive resolver
-    let permissive = build_services(db.clone(), ServiceConfig::default());
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let city = permissive
-        .cities
-        .create_city(
-            &ctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Del City".to_owned(),
-                country: "DL".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    let addr = permissive
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "Del St".to_owned(),
-                postal_code: "22222".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Now use the denying resolver
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-
-    let err = services
-        .addresses
-        .delete_address(&ctx, addr.id)
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden for delete_address, got: {err:?}"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// PDP evaluation failure tests (AuthZResolverError::Internal → DomainError::InternalError)
-// ---------------------------------------------------------------------------
-
-/// When the PDP returns an internal error (e.g., unreachable), the service
-/// must propagate it as `DomainError::InternalError`, not `Forbidden`.
-#[tokio::test]
-async fn pdp_internal_error_returns_internal_for_list_users() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(FailingAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .users
-        .list_users_page(&ctx, &modkit_odata::ODataQuery::default())
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::InternalError),
-        "Expected DomainError::InternalError from PDP failure, got: {err:?}"
-    );
-}
-
-/// PDP internal error on `create_address` → `DomainError::InternalError`.
-#[tokio::test]
-async fn pdp_internal_error_returns_internal_for_create_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "fail@example.com", "Fail User").await;
-
-    // Create a city with a permissive resolver first
-    let permissive = build_services(db.clone(), ServiceConfig::default());
-    let pctx = ctx_allow_tenants(&[tenant_id]);
-
-    let city = permissive
-        .cities
-        .create_city(
-            &pctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Fail City".to_owned(),
-                country: "FL".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Now use the failing resolver
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(FailingAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "Fail St".to_owned(),
-                postal_code: "99999".to_owned(),
-            },
-        )
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::InternalError),
-        "Expected DomainError::InternalError from PDP failure, got: {err:?}"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Additional PDP denial tests (decision=false → EnforcerError::Denied → Forbidden)
-// ---------------------------------------------------------------------------
-
-/// PDP returns `decision=false` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn decision_false_returns_forbidden_for_list_users() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .users
-        .list_users_page(&ctx, &modkit_odata::ODataQuery::default())
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden from decision=false, got: {err:?}"
-    );
-}
-
-/// PDP returns `decision=false` on `create_address` → `DomainError::Forbidden`.
-#[tokio::test]
-async fn decision_false_returns_forbidden_for_create_address() {
-    let db = inmem_db().await;
-    let tenant_id = Uuid::new_v4();
-    let user_id = Uuid::new_v4();
-    let conn = db.conn().unwrap();
-    seed_user(&conn, user_id, tenant_id, "dec@example.com", "Dec User").await;
-
-    // Create a city with a permissive resolver first
-    let permissive = build_services(db.clone(), ServiceConfig::default());
-    let pctx = ctx_allow_tenants(&[tenant_id]);
-
-    let city = permissive
-        .cities
-        .create_city(
-            &pctx,
-            NewCity {
-                id: None,
-                tenant_id,
-                name: "Dec City".to_owned(),
-                country: "DF".to_owned(),
-            },
-        )
-        .await
-        .unwrap();
-
-    // Now use the decision-denied resolver
-    let services = build_services_with_authz(
-        db.clone(),
-        ServiceConfig::default(),
-        Arc::new(DenyAllAuthZResolver),
-    );
-    let ctx = ctx_allow_tenants(&[tenant_id]);
-
-    let err = services
-        .addresses
-        .create_address(
-            &ctx,
-            NewAddress {
-                id: None,
-                tenant_id,
-                user_id,
-                city_id: city.id,
-                street: "Dec St".to_owned(),
-                postal_code: "33333".to_owned(),
-            },
-        )
-        .await
-        .unwrap_err();
-
-    assert!(
-        matches!(err, DomainError::Forbidden),
-        "Expected DomainError::Forbidden from decision=false, got: {err:?}"
-    );
-}
